Reduce size of JS files with code splitting
