+++ app/frontend/containers/EditReviewDialogContainer.js
@@ -12,7 +12,10 @@ import fetchSpots from '../actions/fetchSpots';
 import openPlaceSelectDialog from '../actions/openPlaceSelectDialog';
 import requestMapCenter from '../actions/requestMapCenter';
 import selectSpot from '../actions/selectSpot';
-import { sleep, uploadToStorage, deleteFromStorage, canvasToBlob } from './Utils';
+import uploadToStorage from '../utils/uploadToStorage';
+import deleteFromStorage from '../utils/deleteFromStorage';
+import canvasToBlob from '../utils/canvasToBlob';
+import sleep from '../utils/sleep';
 import I18n from './I18n';
 import openCreateMapDialog from '../actions/openCreateMapDialog';
 import fetchPostableMaps from '../actions/fetchPostableMaps';
