+++ app/frontend/index.js
@@ -26,18 +26,9 @@ import profileReducer from './reducers/profileReducer';
 import persistState from 'redux-localstorage';
 
 import { createLogger } from 'redux-logger';
+import initializeFirebaseApp from './utils/initializeFirebaseApp';
 
-import(/* webpackChunkName: "firebase" */ 'firebase/app').then((firebase) => {
-  const config = {
-    apiKey: process.env.FIREBASE_API_KEY,
-    authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-    projectId: process.env.FIREBASE_PROJECT_ID,
-    databaseURL: process.env.FIREBASE_DB_URL,
-    storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-    messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID
-  };
-  firebase.initializeApp(config);
-});
+initializeFirebaseApp();
 
 const middlewares = [];
