+++ app/frontend/utils/fetchCurrentPosition.js
@@ -1,0 +1,7 @@
+const fetchCurrentPosition = (options = {}) => {
+  return new Promise((resolve, reject) => {
+    navigator.geolocation.getCurrentPosition(resolve, reject, options);
+  });
+};
+
+export default fetchCurrentPosition;
