+++ app/frontend/utils/getFirebase.js
@@ -1,0 +1,5 @@
+const getFirebase = () => {
+  return import(/* webpackChunkName: "firebase" */ 'firebase/app');
+};
+
+export default getFirebase;
