+++ app/frontend/utils/getFirebaseAuth.js
@@ -1,0 +1,5 @@
+const getFirebaseAuth = () => {
+  return import(/* webpackChunkName: "firebase_auth" */ '@firebase/auth');
+};
+
+export default getFirebaseAuth;
