+++ app/frontend/utils/getFirebaseMessaging.js
@@ -1,0 +1,5 @@
+const getFirebaseMessaging = () => {
+  return import(/* webpackChunkName: "firebase_messaging" */ '@firebase/messaging');
+};
+
+export default getFirebaseMessaging;
