+++ app/frontend/utils/getFirebaseStorage.js
@@ -1,0 +1,5 @@
+const getFirebaseStorage = () => {
+  return import(/* webpackChunkName: "firebase_storage" */ '@firebase/storage');
+};
+
+export default getFirebaseStorage;
