+++ app/frontend/utils/getFirebaseUi.js
@@ -1,0 +1,5 @@
+const getFirebaseUi = () => {
+  return import(/* webpackChunkName: "firebase_ui" */ 'firebaseui');
+};
+
+export default getFirebaseUi;
