+++ app/frontend/utils/getFirestore.js
@@ -1,0 +1,5 @@
+const getFirestore = () => {
+  return import(/* webpackChunkName: "firestore" */ '@firebase/firestore');
+};
+
+export default getFirestore;
