+++ app/frontend/utils/initializeFirebaseApp.js
@@ -1,0 +1,17 @@
+import getFirebase from './getFirebase';
+
+const initializeFirebaseApp = async () => {
+  const firebase = await getFirebase();
+
+  const config = {
+    apiKey: process.env.FIREBASE_API_KEY,
+    authDomain: process.env.FIREBASE_AUTH_DOMAIN,
+    projectId: process.env.FIREBASE_PROJECT_ID,
+    databaseURL: process.env.FIREBASE_DB_URL,
+    storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
+    messagingSenderId: process.env.FIREBASE_MESSAGING_SENDER_ID
+  };
+  firebase.initializeApp(config);
+};
+
+export default initializeFirebaseApp;
