+++ app/frontend/utils/sleep.js
@@ -1,0 +1,7 @@
+const sleep = msec => {
+  return new Promise((resolve, reject) => {
+    setTimeout(resolve, msec);
+  });
+};
+
+export default sleep;
